refactor CustomStatusBar into it's own folder
