+++ src/components/StatusBar/index.js
@@ -1,0 +1,2 @@
+import CustomStatusBar from "./CustomStatusBar";
+export default CustomStatusBar;
